# src/main/resources/static/api-docs.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "DataElementHub Rest API",
    "contact": {
      "name": "Medical Informatics Group",
      "url": "http://www.mig-frankfurt.de",
      "email": "vengadeswaran@med.uni-frankfurt.de"
    },
    "description": "REST API Documentation!\n\n  Github repositories:\n\n<https://github.com/mig-frankfurt/dataelementhub.dal>\n\n<https://github.com/mig-frankfurt/dataelementhub.model>\n\n<https://github.com/mig-frankfurt/dataelementhub.rest>\n\n<https://github.com/mig-frankfurt/dataelementhub.gui>\n\nContact:\n\n",
    "version": "1.0.0",
    "license": {
      "url": "https://www.gnu.org/licenses/agpl-3.0.de.html",
      "name": "License: AGPLv3"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8090",
      "description": "Generated server url"
    }
  ],
  "paths": {
    "/element/{urn}/release": {
      "patch": {
        "tags": [
          "Element"
        ],
        "summary": "Release Element by URN.",
        "operationId": "release",
        "parameters": [
          {
            "name": "urn",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "default response",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "security": [
          {
            "basicAuth": []
          }
        ]
      }
    },
    "/element/{urn}": {
      "get": {
        "tags": [
          "Element"
        ],
        "summary": "Get Element by URN.",
        "operationId": "read",
        "parameters": [
          {
            "name": "urn",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Accept-Language",
            "type": "string",
            "required": false,
            "example": "de-DE"
          }
        ],
        "responses": {
          "200": {
            "description": "default response",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "security": [
          {
            "basicAuth": []
          }
        ]
      },
      "put": {
        "tags": [
          "Element"
        ],
        "summary": "Update an Element with one of the models Dataelement, Dataelementgroup or Record",
        "operationId": "update",
        "parameters": [
          {
            "name": "urn",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "default response",
          "content": {
            "*/*": {
              "schema": {
                "oneOf": [
                  {
                    "$ref": "#/components/schemas/DataElement"
                  },
                  {
                    "$ref": "#/components/schemas/DataElementGroup"
                  },
                  {
                    "$ref": "#/components/schemas/Record"
                  }
                ]
              },
              "examples": {
                "DATAELEMENT": {
                  "description": "DATAELEMENT",
                  "value": {
                    "definitions": [
                      {
                        "designation": "Geschlecht",
                        "definition": "Geschlecht des Patienten",
                        "language": "de"
                      },
                      {
                        "designation": "gender",
                        "definition": "patient´s gender",
                        "language": "en"
                      }
                    ],
                    "valueDomain": {
                      "type": "STRING",
                      "text": {
                        "useRegEx": false,
                        "regEx": null,
                        "useMaximumLength": false
                      }
                    },
                    "slots": [
                      {
                        "name": "slot_name",
                        "value": "slot_value"
                      }
                    ],
                    "conceptAssociations": [
                      {
                        "conceptId": 1,
                        "system": "ca_system",
                        "sourceId": 1,
                        "version": "ca_version",
                        "term": "ca_term",
                        "text": "ca_text",
                        "linktype": "undefined",
                        "scopedIdentifierId": 1
                      }
                    ]
                  }
                },
                "DATAELEMENTGROUP": {
                  "description": "DATAELEMENTGROUP",
                  "value": {
                    "definitions": [
                      {
                        "designation": "people",
                        "definition": "smartphone",
                        "language": "de"
                      }
                    ],
                    "members": {
                      "dataElementGroups": [],
                      "dataElements": [
                        "urn:1:dataelement:1:1"
                      ],
                      "records": []
                    },
                    "slots": [
                      {
                        "name": "slot_DEG",
                        "value": "slot_value"
                      }
                    ]
                  }
                },
                "RECORD": {
                  "description": "RECORD",
                  "value": {
                    "definitions": [
                      {
                        "designation": "people",
                        "definition": "smartphone",
                        "language": "de"
                      }
                    ],
                    "members": {
                      "dataElementGroups": [],
                      "dataElements": [
                        "urn:1:dataelement:1:1"
                      ],
                      "records": []
                    },
                    "slots": [
                      {
                        "name": "slot_DEG",
                        "value": "slot_value"
                      }
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "default response",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "security": [
          {
            "basicAuth": []
          }
        ]
      },
      "delete": {
        "tags": [
          "Element"
        ],
        "summary": "Delete Element by URN.",
        "operationId": "delete",
        "parameters": [
          {
            "name": "urn",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "default response",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "security": [
          {
            "basicAuth": []
          }
        ]
      }
    },
    "/element/{urn}/valuedomain": {
      "get": {
        "tags": [
          "Element"
        ],
        "summary": "Get ValueDomain from a specified Element.",
        "operationId": "release",
        "parameters": [
          {
            "name": "urn",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "default response",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "security": [
          {
            "basicAuth": []
          }
        ]
      }
    },
    "/element/{urn}/definitions": {
      "get": {
        "tags": [
          "Element"
        ],
        "summary": "Get Definitions from a specified Element.",
        "operationId": "release",
        "parameters": [
          {
            "name": "urn",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "default response",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "security": [
          {
            "basicAuth": []
          }
        ]
      }
    },
    "/element/{urn}/slots": {
      "get": {
        "tags": [
          "Element"
        ],
        "summary": "Get Slots from a specified Element.",
        "operationId": "release",
        "parameters": [
          {
            "name": "urn",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "default response",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "security": [
          {
            "basicAuth": []
          }
        ]
      }
    },
    "/element/{urn}/validation": {
      "get": {
        "tags": [
          "Element"
        ],
        "summary": "Get Validation from a specified Element.",
        "operationId": "release",
        "parameters": [
          {
            "name": "urn",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "default response",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "security": [
          {
            "basicAuth": []
          }
        ]
      }
    },
    "/element/{urn}/identification": {
      "get": {
        "tags": [
          "Element"
        ],
        "summary": "Get Identification from a specified Element.",
        "operationId": "release",
        "parameters": [
          {
            "name": "urn",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "default response",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "security": [
          {
            "basicAuth": []
          }
        ]
      }
    },
    "/element/{urn}/conceptassociation": {
      "get": {
        "tags": [
          "Element"
        ],
        "summary": "Get ConceptsAssociations from a specified Element.",
        "operationId": "release",
        "parameters": [
          {
            "name": "urn",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "default response",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "security": [
          {
            "basicAuth": []
          }
        ]
      }
    },
    "/element": {
      "post": {
        "tags": [
          "Element"
        ],
        "summary": "Create an Element with one of the models Dataelement, Dataelementgroup or Record",
        "operationId": "create",
        "requestBody": {
          "description": "default response",
          "content": {
            "application/json": {
              "schema": {
                "oneOf": [
                  {
                    "$ref": "#/components/schemas/DataElement"
                  },
                  {
                    "$ref": "#/components/schemas/DataElementGroup"
                  },
                  {
                    "$ref": "#/components/schemas/Record"
                  }
                ]
              },
              "examples": {
                "DATAELEMENT": {
                  "description": "DATAELEMENT",
                  "value": {
                    "identification": {
                      "elementType": "DATAELEMENT",
                      "namespaceId": 1,
                      "namespaceUrn": "urn:1:namespace:1:1",
                      "status": "RELEASED"
                    },
                    "definitions": [
                      {
                        "designation": "Geschlecht",
                        "definition": "Geschlecht des Patienten",
                        "language": "de"
                      },
                      {
                        "designation": "gender",
                        "definition": "patient´s gender",
                        "language": "en"
                      }
                    ],
                    "valueDomain": {
                      "type": "STRING",
                      "text": {
                        "useRegEx": false,
                        "regEx": null,
                        "useMaximumLength": false
                      }
                    },
                    "slots": [
                      {
                        "name": "slot_name",
                        "value": "slot_value"
                      }
                    ],
                    "conceptAssociations": [
                      {
                        "conceptId": 1,
                        "system": "ca_system",
                        "sourceId": 1,
                        "version": "ca_version",
                        "term": "ca_term",
                        "text": "ca_text",
                        "linktype": "undefined",
                        "scopedIdentifierId": 1
                      }
                    ]
                  }
                },
                "DATAELEMENTGROUP": {
                  "description": "DATAELEMENTGROUP",
                  "value": {
                    "identification": {
                      "elementType": "DATAELEMENTGROUP",
                      "namespaceId": 1,
                      "status": "RELEASED"
                    },
                    "definitions": [
                      {
                        "designation": "people",
                        "definition": "smartphone",
                        "language": "de"
                      }
                    ],
                    "members": {
                      "dataElementGroups": [],
                      "dataElements": [
                        "urn:1:dataelement:1:1"
                      ],
                      "records": []
                    },
                    "slots": [
                      {
                        "name": "slot_DEG",
                        "value": "slot_value"
                      }
                    ]
                  }
                },
                "RECORD": {
                  "description": "RECORD",
                  "value": {
                    "identification": {
                      "elementType": "RECORD",
                      "namespaceId": 1,
                      "status": "RELEASED"
                    },
                    "definitions": [
                      {
                        "designation": "people",
                        "definition": "smartphone",
                        "language": "de"
                      }
                    ],
                    "members": {
                      "dataElementGroups": [],
                      "dataElements": [
                        "urn:1:dataelement:1:1"
                      ],
                      "records": []
                    },
                    "slots": [
                      {
                        "name": "slot_DEG",
                        "value": "slot_value"
                      }
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "CREATED",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "security": [
          {
            "basicAuth": []
          }
        ]
      }
    },
    "/namespaces/{namespaceId}/release": {
      "patch": {
        "tags": [
          "Namespace"
        ],
        "summary": "Release Namespace by id.",
        "operationId": "release_1",
        "parameters": [
          {
            "name": "namespaceId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "default response",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "security": [
          {
            "basicAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DataElement": {
        "type": "object",
        "properties": {
          "identification": {
            "$ref": "#/components/schemas/Identification"
          },
          "definitions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Definition"
            }
          },
          "slots": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Slot"
            }
          },
          "valueDomain": {
            "$ref": "#/components/schemas/ValueDomain"
          },
          "conceptAssociations": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ConceptAssociation"
            }
          }
        }
      },
...
      "Definition": {
        "type": "object",
        "properties": {
          "designation": {
            "type": "string"
          },
          "definition": {
            "type": "string"
          },
          "language": {
            "type": "string"
          }
        }
      },
...
      "Slot": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "value": {
            "type": "string"
          }
        }
      },
...
      "DataElementGroup": {
        "type": "object",
        "properties": {
          "identification": {
            "$ref": "#/components/schemas/Identification"
          },
          "definitions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Definition"
            }
          },
          "slots": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Slot"
            }
          },
          "members": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Member"
            }
          }
        }
      },
...
      "Record": {
        "type": "object",
        "properties": {
          "identification": {
            "$ref": "#/components/schemas/Identification"
          },
          "definitions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Definition"
            }
          },
          "slots": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Slot"
            }
          },
          "members": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Member"
            }
          }
        }
      },
...
    },
    "securitySchemes": {
      "basicAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}